+++ packages/create-app/templates/default-app/packages/client/src/Root.tsx
@@ -1,12 +1,12 @@
 import React, { useMemo } from 'react'
-import { determineTheme } from './utils/themes'
+import { defaultTheme } from './utils/themes'
 import loadConfig from './ConfigLoader'
 import { BrowserRouter } from 'react-router-dom'
 import { CssBaseline, ThemeProvider } from '@material-ui/core'
 import { App } from './App'
 
 export function Root() {
-  const theme = useMemo(() => determineTheme(), [])
+  const theme = useMemo(() => defaultTheme(), [])
   const config = useMemo(() => loadConfig(), [])
 
   return (