+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cloud-carbon-footprint/azure
 
+## 0.10.0
+
+### Minor Changes
+
+- dcc33152: Adds support for Azure in lookup table
+
+## 0.9.2
+
+### Patch Changes
+
+- 84e2d1e2: adds support for missing usage types
+
 ## 0.9.1
 
 ### Patch Changes